# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "思维模型平台 - 基于结构化思维模型和AI的复杂问题深度思考平台",
        "title": "Thinking Models Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@thinkingmodels.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:2500",
    "basePath": "/api/v1",
    "paths": {
...
        "/thinking/model": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建一个新的思维模型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "思维模型"
                ],
                "summary": "创建思维模型",
                "parameters": [
                    {
                        "description": "创建思维模型请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/thinkingModels_domain_thinking_model.CreateModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/thinkingModels_domain_thinking_model.ModelDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/thinking/model/list": {
            "get": {
                "description": "分页查询思维模型列表，支持多种筛选条件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "思维模型"
                ],
                "summary": "获取思维模型列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认10",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称模糊查询",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分类ID",
                        "name": "categoryId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "难度：1-简单，2-中等，3-困难",
                        "name": "difficulty",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否免费",
                        "name": "isFree",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/thinkingModels_domain_thinking_model.ListModelResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/thinking/model/publish": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "将思维模型发布到市场",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "思维模型"
                ],
                "summary": "发布思维模型",
                "parameters": [
                    {
                        "description": "发布请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/thinkingModels_domain_thinking_model.PublishModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "发布成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/thinkingModels_domain_thinking_model.ModelInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/thinking/model/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据ID获取思维模型详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "思维模型"
                ],
                "summary": "获取思维模型详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "模型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/thinkingModels_domain_thinking_model.ModelDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "模型不存在",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "trace_id": {}
            }
        },
...
        "thinkingModels_domain_thinking_model.CreateModel": {
            "type": "object",
            "required": [
                "categoryId",
                "code",
                "name"
            ],
            "properties": {
                "authorName": {
                    "type": "string",
                    "maxLength": 100
                },
                "categoryId": {
                    "type": "integer"
                },
                "code": {
                    "type": "string",
                    "maxLength": 50
                },
                "content": {
                    "type": "string"
                },
                "coverImage": {
                    "type": "string",
                    "maxLength": 500
                },
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "difficulty": {
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3
                    ]
                },
                "estimatedTime": {
                    "type": "integer",
                    "minimum": 1
                },
                "icon": {
                    "type": "string",
                    "maxLength": 200
                },
                "isOfficial": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100
                },
                "overview": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                },
                "version": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
        "thinkingModels_domain_thinking_model.ListModelResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/thinkingModels_domain_thinking_model.ModelInfo"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "thinkingModels_domain_thinking_model.ModelAuthor": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "thinkingModels_domain_thinking_model.ModelDetail": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/thinkingModels_domain_thinking_model.ModelAuthor"
                },
                "categoryId": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "coverImage": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "integer"
                },
                "estimatedTime": {
                    "type": "integer"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isFree": {
                    "type": "boolean"
                },
                "isOfficial": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "publishTime": {
                    "type": "string"
                },
                "sourceModelId": {
                    "type": "integer"
                },
                "stats": {
                    "$ref": "#/definitions/thinkingModels_domain_thinking_model.ModelStats"
                },
                "status": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "thinkingModels_domain_thinking_model.ModelInfo": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/thinkingModels_domain_thinking_model.ModelAuthor"
                },
                "categoryId": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "coverImage": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "integer"
                },
                "estimatedTime": {
                    "type": "integer"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isFree": {
                    "type": "boolean"
                },
                "isOfficial": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "publishTime": {
                    "type": "string"
                },
                "sourceModelId": {
                    "type": "integer"
                },
                "stats": {
                    "$ref": "#/definitions/thinkingModels_domain_thinking_model.ModelStats"
                },
                "status": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "thinkingModels_domain_thinking_model.ModelStats": {
            "type": "object",
            "properties": {
                "adoptCount": {
                    "type": "integer"
                },
                "commentCount": {
                    "type": "integer"
                },
                "likeCount": {
                    "type": "integer"
                },
                "usageCount": {
                    "type": "integer"
                }
            }
        },
        "thinkingModels_domain_thinking_model.PublishModel": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT token认证，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}